+++ components/index.js
@@ -1,395 +1,15 @@
-import {  keys, keysObj, lang, specialKeys } from './keys.js'
+import {  drawKeyboard } from './keyboard.js';
+import {  curLang, setLang } from './langSwitcher.js';
+import {  curMode } from './modeSwitcher.js';
+import {  focusTo } from './utils.js';
+import {  handlerMouseDown, handlerMouseUp, handlerKeydown, handlerKeyUp } from './handlers.js';
 
-let curLang = window.localStorage.getItem('lang') ? window.localStorage.getItem('lang') : 'rus';
-window.localStorage.setItem('lang', curLang);
-let curMode = 'caseDown';
-let pressedKey;
+setLang(curLang, curMode.mode);
+drawKeyboard();
 
-const body = document.querySelector('.body');
-
-// Page
-const page = document.createElement('div');
-page.className = 'page';
-body.prepend(page);
-
-// Header
-const header = document.createElement('header');
-header.className = 'header';
-page.prepend(header);
-
-// Heading
-const heading = document.createElement('h1');
-heading.className = 'header__heading';
-heading.textContent = 'Virtual keyboard';
-header.prepend(heading);
-
-// Main
-const main = document.createElement('main');
-main.className = 'main';
-header.after(main);
-
-// Textarea
-const textarea = document.createElement('textarea');
-textarea.className = 'textarea';
-textarea.id = 'textarea';
-textarea.setAttribute('cols',50);
-textarea.setAttribute('rows',5);
-main.append(textarea);
-
-// Keyboard
-const keyboard = document.createElement('div');
-keyboard.className = 'keyboard';
-keyboard.id = 'keyboard';
-main.append(keyboard);
-
-for (let i = 0; i < 5; i++) {
-  const keyboardRow = document.createElement('div');
-  keyboardRow.className = 'keyboard__row';
-  keyboard.append(keyboardRow);
-
-  for (let j = 0; j < keys[i].length; j++) {
-    // Keyboard key
-    const keyboardKey = document.createElement('div');
-    keyboardKey.className = 'keyboard__key';
-    keyboardKey.classList.add(keys[i][j]);
-    keyboardRow.append(keyboardKey);
-    // Langs
-    for (let k = 0; k < lang.length; k++) {
-      const spanLang = document.createElement('span');
-      const isHidden = lang[k] === curLang ? '' : ' hidden';
-      spanLang.className = lang[k] + isHidden;
-      keyboardKey.append(spanLang);
-      // caseDown
-      const caseDown = document.createElement('span');
-      caseDown.className = 'caseDown' + isHidden;
-      caseDown.textContent = keysObj[keys[i][j]][lang[k]].caseDown;
-      spanLang.append(caseDown);
-      // caseUp
-      const caseUp = document.createElement('span');
-      caseUp.className = 'caseUp hidden';
-      caseUp.textContent = keysObj[keys[i][j]][lang[k]].caseUp;
-      spanLang.append(caseUp);
-      // caseUp
-      const caps = document.createElement('span');
-      caps.className = 'caps hidden';
-      caps.textContent = keysObj[keys[i][j]][lang[k]].caps;
-      spanLang.append(caps);
-      // shiftCaps
-      const shiftCaps = document.createElement('span');
-      shiftCaps.className = 'shiftCaps hidden';
-      shiftCaps.textContent = keysObj[keys[i][j]][lang[k]].shiftCaps;
-      spanLang.append(shiftCaps);
-    }
-  }
-}
-
-// Footer
-const footer = document.createElement('footer');
-footer.className = 'footer';
-main.after(footer);
-
-// Footer text
-const p1 = document.createElement('p');
-const p2= document.createElement('p');
-p1.className = 'footer__text';
-p2.className = 'footer__text';
-footer.append(p1);
-footer.append(p2);
-p1.textContent = 'Для смены языка используйте сочетание Left Ctrl + Alt';
-p2.textContent = 'Клавиатура создана в операционной системе Windows';
-
-const capsAll = document.querySelectorAll('.caps');
-const caseDownAll = document.querySelectorAll('.caseDown');
-const caseUpAll = document.querySelectorAll('.caseUp');
-const shiftCapsAll = document.querySelectorAll('.shiftCaps');
-
-const setCaps = () => {
-  capsAll.forEach(item => item.classList.remove('hidden'));
-  caseDownAll.forEach(item => item.classList.add('hidden'));
-  caseUpAll.forEach(item => item.classList.add('hidden'));
-  shiftCapsAll.forEach(item => item.classList.add('hidden'));
-}
-
-const setCaseDown = () => {
-  capsAll.forEach(item => item.classList.add('hidden'));
-  caseDownAll.forEach(item => item.classList.remove('hidden'));
-  caseUpAll.forEach(item => item.classList.add('hidden'));
-  shiftCapsAll.forEach(item => item.classList.add('hidden'));
-}
-
-const setCaseUp = () => {
-  capsAll.forEach(item => item.classList.add('hidden'));
-  caseDownAll.forEach(item => item.classList.add('hidden'));
-  caseUpAll.forEach(item => item.classList.remove('hidden'));
-  shiftCapsAll.forEach(item => item.classList.add('hidden'));
-}
-
-const setShiftCaps = () => {
-  capsAll.forEach(item => item.classList.add('hidden'));
-  caseDownAll.forEach(item => item.classList.add('hidden'));
-  caseUpAll.forEach(item => item.classList.add('hidden'));
-  shiftCapsAll.forEach(item => item.classList.remove('hidden'));
-}
-
-const rusSpans = document.querySelectorAll('.rus');
-const engSpans = document.querySelectorAll('.eng');
-
-const setLang = (lang) => {
-  if (lang == 'rus') {
-    engSpans.forEach(item => item.classList.add('hidden'));
-    rusSpans.forEach(item => item.classList.remove('hidden'));
-  }
-  if (lang == 'eng') {
-    engSpans.forEach(item => item.classList.remove('hidden'));
-    rusSpans.forEach(item => item.classList.add('hidden'));
-  }
-  switch (curMode) {
-    case 'caseDown':
-      setCaseDown();
-      break;
-    case 'caseUp':
-      setCaseUp();
-      break;
-    case 'caps':
-      setCaps();
-      break;
-    case 'shiftCaps':
-      setShiftCaps();
-      break;
-  }
-  curLang = lang;
-  window.localStorage.setItem('lang', curLang);
-}
-
-document.addEventListener('mouseover', (evt) => {
-  textarea.focus();
-});
-
-document.addEventListener('click', (evt) => {
-  textarea.focus();
-});
-
-document.addEventListener('mousedown', (evt) => {
-  textarea.focus();
-  if (evt.target.closest('.keyboard__key')) {
-
-    const key = evt.target.closest('.keyboard__key');
-    pressedKey = key;
-    if (pressedKey.classList.contains('CapsLock') && curMode !== 'caseUp') {
-      key.classList.toggle('pressed');
-      switch (curMode) {
-        case 'caps':
-          curMode = 'caseDown';
-          setCaseDown();
-          break;
-        case 'shiftCaps':
-          curMode = 'caseUp';
-          setCaseUp();
-          break;
-        default:
-          curMode = 'caps';
-          setCaps();
-      }
-    } else if ((pressedKey.classList.contains('ShiftLeft') || pressedKey.classList.contains('ShiftRight')) && curMode !== 'caps') {
-
-      key.classList.toggle('pressed');
-
-        switch (curMode) {
-          case 'caseUp':
-            curMode = 'caseDown';
-            setCaseDown();
-            break;
-          case 'shiftCaps':
-            curMode = 'caps';
-            setCaps();
-            break;
-          default:
-            curMode = 'caseUp';
-            setCaseUp();
-        }
-
-
-    } else if ((pressedKey.classList.contains('CapsLock') && curMode == 'caseUp') ||
-    ((pressedKey.classList.contains('ShiftLeft') || pressedKey.classList.contains('ShiftRight')) && curMode == 'caps')) {
-      console.log(777);
-      key.classList.toggle('pressed');
-      curMode = 'shiftCaps';
-      setShiftCaps();
-    } else if (pressedKey.classList.contains('ToggleLang')) {
-      curLang === 'rus' ? setLang('eng') : setLang('rus');
-      key.classList.add('pressed');
-    } else {
-      key.classList.add('pressed');
-    }
-
-    insertSymbols(key);
-
-
-  }
-});
-
-
-function insertSymbols(key) {
-    const charCode = key.classList[1];
-    const char = keysObj[charCode][curLang][curMode];
-    let value = textarea.value;
-    let start;
-    let end;
-    let replaceLeft;
-    let replaceRight;
-    if (textarea.selectionStart == textarea.selectionEnd) {
-      start = end = textarea.selectionStart;
-      replaceLeft = start - 1;
-      replaceRight = end + 1;
-    } else {
-      start = textarea.selectionStart;
-      end = textarea.selectionEnd;
-      replaceLeft = start;
-      replaceRight = end;
-    }
-    if (specialKeys.includes(charCode)) {
-      switch (charCode) {
-        case 'Backspace':
-          if (start >= 0 && start <= value.length) {
-            value = value.slice(0, replaceLeft) + value.slice(end, value.length),
-            textarea.value = value;
-            textarea.selectionStart = replaceLeft;
-            textarea.selectionEnd = replaceLeft;
-          };
-          break;
-        case 'Delete':
-          if (start >= 0 && start <= value.length - 1) {
-            value = value.slice(0, start) + value.slice(replaceRight, value.length);
-            textarea.value = value;
-            textarea.selectionStart = start;
-            textarea.selectionEnd = start;
-          };
-          break;
-        case 'Enter':
-          if (start >= 0 && start <= value.length) {
-            value = value.slice(0, start) + '\n' + value.slice(end, value.length);
-            textarea.value = value;
-            textarea.selectionStart = start + 1;
-            textarea.selectionEnd = start + 1;
-          } else {
-            textarea.value += '\n';
-          };
-          break;
-        case 'Tab':
-          if (start >= 0 && start <= value.length) {
-            value = value.slice(0, start) + '    ' + value.slice(end, value.length);
-            textarea.value = value;
-            textarea.selectionStart = start + 4;
-            textarea.selectionEnd = start + 4;
-          } else {
-            textarea.value += '    ';
-          };
-          break;
-        case 'ShiftLeft':
-          //console.log('ShiftLeft');
-          break;
-        case 'ShiftRight':
-          //console.log('ShiftRight');
-          break;
-        case 'CapsLock':
-          //console.log('CapsLock');
-          break;
-      }
-    } else {
-      if (start >= 0 && start <= value.length) {
-        textarea.value = value.slice(0, start) + char + value.slice(end, value.length);
-        textarea.selectionStart = start + char.length;
-        textarea.selectionEnd = start + char.length;
-      } else {
-        textarea.value += char;
-      }
-    }
-}
-
-document.addEventListener('mouseup', (evt) => {
-  if (typeof pressedKey === 'object' && pressedKey.className.includes('pressed')) {
-    if (!pressedKey.classList.contains('CapsLock') && !pressedKey.classList.contains('ShiftLeft') && !pressedKey.classList.contains('ShiftRight')) {
-      pressedKey.classList.remove('pressed');
-    }
-  }
-});
-
-document.addEventListener('keydown', (evt) => {
-
-  if (evt.ctrlKey && evt.key === 'Alt')  {
-    curLang === 'rus' ? setLang('eng') : setLang('rus');
-  }
-
-  evt.preventDefault();
-  pressedKey = evt.code;
-  const key = document.querySelector('.' + evt.code);
-
-  //console.log(evt.code);
-  if (pressedKey == 'CapsLock' && curMode !== 'caseUp') {
-    key.classList.toggle('pressed');
-    switch (curMode) {
-      case 'caps':
-        curMode = 'caseDown';
-        setCaseDown();
-        break;
-      case 'shiftCaps':
-        curMode = 'caseUp';
-        setCaseUp();
-        break;
-      default:
-        curMode = 'caps';
-        setCaps();
-    }
-
-  } else if ((pressedKey == 'ShiftLeft' || pressedKey == 'ShiftRight') && curMode !== 'caps') {
-    key.classList.toggle('pressed');
-    switch (curMode) {
-      case 'caseUp':
-        curMode = 'caseDown';
-        setCaseDown();
-        break;
-      case 'shiftCaps':
-        curMode = 'caps';
-        setCaps();
-        break;
-      default:
-        curMode = 'caseUp';
-        setCaseUp();
-    }
-  } else if ((pressedKey == 'CapsLock' && curMode == 'caseUp') ||
-  ((pressedKey == 'ShiftLeft' || pressedKey == 'ShiftRight') && curMode == 'caps')) {
-    console.log(777);
-    key.classList.toggle('pressed');
-    curMode = 'shiftCaps';
-    setShiftCaps();
-  } else {
-    key.classList.add('pressed');
-  }
-
-  insertSymbols(key);
-});
-
-document.addEventListener('keyup', (evt) => {
-  pressedKey = evt.code;
-  const key = document.querySelector('.' + evt.code);
-
-  if (pressedKey !== 'CapsLock' && pressedKey !== 'ShiftLeft' && pressedKey !== 'ShiftRight') {
-    key.classList.remove('pressed');
-  }
-  if (pressedKey == 'ShiftLeft' || pressedKey == 'ShiftRight') {
-    key.classList.remove('pressed');
-    switch (curMode) {
-      case 'caseUp':
-        curMode = 'caseDown';
-        setCaseDown();
-        break;
-      case 'shiftCaps':
-        curMode = 'caps';
-        setCaps();
-        break;
-      default:
-        curMode = 'caseUp';
-        setCaseUp();
-    }
-  }
-});
+document.addEventListener('mouseover', evt => focusTo(textarea));
+document.addEventListener('click', evt => focusTo(textarea));
+document.addEventListener('mousedown', handlerMouseDown);
+document.addEventListener('mouseup', handlerMouseUp);
+document.addEventListener('keydown', handlerKeydown);
+document.addEventListener('keyup', handlerKeyUp);
