+++ index.html
@@ -8,7 +8,7 @@
     <link rel="stylesheet" href="./pages/index.css">
     <link rel="icon" href="./images/favicon.ico">
 </head>
-<body class="body">
+<body>
 
 </body>
 </html>
